server/requirements.txt
# Zeni Server Dependencies
# Real-time Voice AI System

# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
websockets==12.0

# AI/ML
numpy==1.26.2
soundfile==0.12.1

# HTTP Client
aiohttp==3.9.1

# Google Cloud Services
google-cloud-speech==2.29.0       # ASR - Speech-to-Text
google-cloud-texttospeech>=2.34.0  # TTS - Text-to-Speech (streaming with Gemini)
google-genai>=0.4.0               # Gemini TTS - Generative AI

# RAG - Vector Search for FAQ
chromadb==0.5.0                    # Vector database (local)
sentence-transformers==3.0.0       # Multilingual embeddings (e5-small on MPS)
torch>=2.0.0                       # Required for sentence-transformers

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
PyYAML==6.0.1

# Async utilities
uvloop==0.19.0

# Logging
structlog==24.1.0

# File uploads
python-multipart>=0.0.6

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.23.2
